Refactor scheduled jobs.
